Automate PyPI Release
These creds are just for demonstration purposes. If there is an
organizational PyPI account to use or if one of the listed Baseplate
maintainers wants to supply theirs, let's talk!

The hope here is to do smaller incremental releases and publish them to
PyPI. This would allow Reddit projects that are using wheels to make use
of new functionality and bug fixes before doing a fleetwide release.

@spladug, any help you could lend with the credentials would be much
appreciated.